requirements.txt
./custom-libs/django-allauth-blocklight
aiohttp==3.6.2
amqp==2.5.2
anyjson==0.3.3
async-timeout==3.0.1
argh==0.26.2
attrs==20.2.0
authclient==1.0
Babel==2.7.0
beautifulsoup4==4.8.0
billiard==3.6.3.0
cachetools==3.1.1
celery==4.4.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==36.0.2
defusedxml==0.6.0
Django==3.2
django-appconf==1.0.3
django-autocomplete-light==3.3.2
django-avatar==5.0.0
django-compat==1.0.15
django-extensions==2.2.9
django-filer==2.0.2
django-filter==2.4
django-hijack==2.1.10
django-invitations==1.9.3
django-nose==1.4.5
django-polymorphic==3.0.0
django-qsstats-magic==1.0.0
django-reset-migrations==0.3.1
django-smtp-ssl==1.0
django-treebeard==4.3
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.11.1
draftjs-exporter==2.1.7
easy-thumbnails==2.7.1
enum-compat==0.0.3
flower==0.9.3
google-api-python-client==1.6.5
google-auth==1.6.3
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.14.0
idna==2.10
idna-ssl==1.1.0
intuit-oauth==1.2.3
jsonpatch==1.24
jsonpointer==2.0
mailchimp-marketing==3.0.12
mailchimp3==3.0.14
mixpanel==4.7.0
mysqlclient==1.4.6
ndg-httpsclient==0.5.1
netifaces==0.10.7
nose==1.3.7
numpy==1.17.2
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==6.2.0
plus==1994.12.4
pyactiveresource==2.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyshorteners==1.0.1
python-dateutil==2.7.3
python-geoip==1.2
python-quickbooks==0.8.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
regex==2018.8.29
requests==2.25.0
requests-oauthlib==1.2.0
rsa==4.0
ShopifyAPI==10.0.0
six==1.15.0
stripe==2.40.0
torchfile==0.1.0
tornado==5.1.1
tweepy==3.9.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.26.1
visdom==0.1.8.9
wagtail==2.13
watchdog==2.1.5
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
Willow==1.4
